==> jupyter_notebooks/01 - DataCollection.ipynb <==
# Data Collection

## Objectives

*   Fetch data from Kaggle competition and save as raw data
*   Get a data summary
*   Inspect the shape of the data
*   Output the data and save in outputs/datasets/collection


## Inputs

* JSON file from Kaggle & Authentication Token

## Outputs

* Generated Dataset saved as outputs/datasets/collection/HousePriceRegression.csv

## Additional Comments


* This data set does not contain any pricate or sensitive information. As such, the data set is committed to a repository which would usually not be the case when data needs to be protected.



---

# Change Working Directory

In order to install Kaggle and fetch the data from Kaggle, the current working directory needs to be changed to the root level of the workspace.

import os

# Get the current working directory (cwd)
cwd = os.getcwd()
print(f"[*] Previous working directory: {cwd}")

# Make the parent of the cwd the new cwd
os.chdir(os.path.dirname(cwd))
cwd = os.getcwd()
print(f"[*] Updated current working directory: {cwd}")

---

# Fetch Data From Kaggle

In order to fetch data from Kaggle using the Kaggle API, Kaggle needs to installed in the workspace. An access token can be obtained using instructions at the [Kaggle API repository](https://github.com/Kaggle/kaggle-api).

> If a cloud-based development environment is being used (e.g. Gitpod), the `kaggle.json` credentials file which gets downloaded to your local machine when creating a key should be uploaded to the remote workspace.

**NB** Your Kaggle API key is private and should not be shared.

# Install Kaggle
! pip install kaggle==1.5.12

# Set up token variable
import os
os.environ['KAGGLE_CONFIG_DIR'] = os.getcwd()

# Set read/write permission for file owner only
! chmod 600 kaggle.json

Download the data to destination folder and unzip it.

# Define data source path
kaggle_dataset_path = "house-prices-advanced-regression-techniques"

# Define destination folder
destination_folder = "inputs/datasets/raw"

#Download the data
! kaggle competitions download -c {kaggle_dataset_path} -p {destination_folder}

# Unzip downloaded file and remove the zip file
! unzip {destination_folder}/*.zip -d {destination_folder} \
  && rm {destination_folder}/*.zip

# Inspect Data

The data set downloaded in the previous section contains data that has already been split into training and test sets.

In this section, we'll:

- import the data into a DataFrame
- get information about the DataFrame
- get the data types for different label colums
- look for missing values

# Import train.csv into a DataFrame and inspect the first 5 entries
import pandas as pd
df = pd.read_csv("inputs/datasets/raw/train.csv")
df.head()

Get a summary of the df DataFrame

df.info()

DataFrame shape

- Rows: 1460
- Colums: 81

df.shape

---
