==> jupyter_notebooks/03 - DataCleaning.ipynb <==
# Data Cleaning

## 1. Objectives

- Evaluate missing data
- Clean data
- Export cleaned data for use in modeling

## 2. Inputs

- /workspace/house-price-regression/inputs/datasets/raw/train.csv

## 3. Outputs

- Clean data sets saved in /workspace/house-price-regression/outputs/datasets/cleaned

## 4. Imports

import pandas as pd
import numpy as np

## 5. Load data

import os

# Get the current working directory (cwd)
cwd = os.getcwd()
print(f"[*] Previous working directory: {cwd}")

# Make the parent of the cwd the new cwd
os.chdir(os.path.dirname(cwd))
cwd = os.getcwd()
print(f"[*] Updated current working directory: {cwd}")

# Load the data
df_train = pd.read_csv("inputs/datasets/raw/train.csv")
df_test = pd.read_csv("inputs/datasets/raw/test.csv")

## 6. Create DataFrames with selected feature variables only

- OverallQual: Overall Quality
- YearBuilt: Year Built
- YearRemodAdd: Remodel Date
- TotalBsmtSF: Total Basement Square Footage
- 1stFlrSF: Total First Floor Square Footage
- GrLivArea: Above Grade (Ground) Living Area Square Footage
- FullBath: Number of Full Bathrooms
- TotRmsAbvGrd: Number of Rooms above Grade
- GarageCars: Garage (Number of Cars)
- GarageArea: Garage Area Square Footage

# Selected features

selected_features_train = ['OverallQual', 'YearBuilt',
                           'YearRemodAdd', 'TotalBsmtSF',
                           '1stFlrSF', 'GrLivArea',
                           'FullBath', 'TotRmsAbvGrd',
                           'GarageCars', 'GarageArea',
                           'SalePrice']

selected_features_test = ['OverallQual', 'YearBuilt',
                     'YearRemodAdd', 'TotalBsmtSF',
                     '1stFlrSF', 'GrLivArea',
                     'FullBath', 'TotRmsAbvGrd',
                     'GarageCars', 'GarageArea']
df_train_selected = df_train[selected_features_train]
df_test_selected = df_test[selected_features_test]

df_test_selected.head()

df_train_selected.head()

## 7. Check for Missing Data

### 7.1 Training data set

# Create function to check for missing values
def check_for_missing_values(df):
    missing_total = df.isnull().sum().sort_values(ascending=False)
    missing_percentage = ((df.isnull().sum()/df.isnull().count()) * 100).sort_values(ascending=False)
    missing_table = pd.concat([missing_total,
                               missing_percentage],
                               axis=1,
                               keys=['Total missing values', 'Percentage'])
    return missing_table

# Check missing values in training data set
df_training_selected_missing_table = check_for_missing_values(df_train_selected)
df_training_selected_missing_table

### 7.2 Test data set

# Check missing values in training data set
df_test_selected_missing_table = check_for_missing_values(df_test_selected)
df_test_selected_missing_table

### 7.3 Missing Values Conclusion

It seems like the training data set contains no missing values for our selected features. The test data set, however, does! Only one value is missing for each of TotalBsmtSF, GarageCars and GarageArea which equates to roughly 0.069% of the test data set. This value is negligible when the size of the data set is considered. As such, these rows can be dropped from the test DataFrame without any reason for concern. 

## 8. Drop Rows With Missing Values

As mentioned in section 7.3, only the test data set contains missing values. As such only the test data set will be targeted for cleaning.

# Check length of test data set before removing missing values
len(df_test_selected)

# Check length of test data set after removing missing values
df_test_selected_clean = df_test_selected.dropna(axis=0)
len(df_test_selected_clean)


# Re-check missing values in training data set after removing missing values
df_test_selected_missing_table = check_for_missing_values(df_test_selected_clean)
df_test_selected_missing_table

## 9. Output Clean Data Sets as CSV files

# Create outputs/datasets/collection folder
import os
try:
  os.makedirs(name='outputs/datasets/cleaned')
except Exception as e:
  print(e)

# Export training data
df_train_selected.to_csv("outputs/datasets/cleaned/TrainSetCleaned.csv", index=False)

# Export test data
df_test_selected_clean.to_csv("outputs/datasets/cleaned/TestSetCleaned.csv", index=False)
